requirements.txt
fastapi==0.109.1
pytest==7.1.3
pytest-asyncio==0.20.3
python-socketio==5.11.4
uvicorn==0.32.1
pydantic~=1.10.4
